refactor: preload script move
